Allow area hiding in contiguous instanced areas
- Handle mistakes in area hiding more gracefully, and log an error.
- Allow world lights to spawn for all copies of the chunk in instances.

+++ src/main/java/rs117/hd/overlays/LightGizmoOverlay.java
@@ -280,7 +280,7 @@ public class LightGizmoOverlay extends Overlay implements MouseListener, KeyList
 					} else {
 						// Shift the position with mouse movement
 						for (int j = 0; j < 2; j++)
-							point[j] += (float) mouseDelta[j];
+							point[j] += mouseDelta[j];
 					}
 
 					if (numFrozenAxes == 0) { // restrict to same depth plane
